requirements.txt
websockets
ocpp
aioconsole
pyyaml
flask
requests
click
cryptography
mitmproxy
